+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - fix bug in history window
 - fix bug in addon loading
+- fix bug in alt ep/gp syncing
 
 # 0.12.2